requirements.txt
# Основная зависимость для MCP сервера
fastmcp>=0.1.0

# Дополнительные зависимости для работы с типами и JSON
typing-extensions>=4.0.0

# Зависимости для OpenRouter клиента
aiohttp>=3.8.0
python-dotenv>=1.0.0 